+++ sdk-core/src/main/java/com/sportradar/unifiedodds/sdk/caching/impl/ci/TournamentStageCIImpl.java
@@ -327,7 +327,7 @@ class TournamentStageCIImpl implements StageCI {
      */
     @Override
     public EventStatus getEventStatus() {
-        return null;
+        return EventStatus.Unknown;
     }
 
     /**