+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 ### Unified Feed SDK 2.x changelog
 
+**2.0.23 (2019-04-18)**
+* exposed Round.getGroupId()
+* improved handling of SportEventStatus updates
+* fix: incorrect message validation
+* fix: initial value for eventStatus
+
+**2.0.22 (2019-04-08)**
+* added support for non-cached fixture endpoint
+* improved fetching logic for the summary endpoint
+* fix: handling pre:outcometext and simpleteam ids in cache
+* fix: null check for names in Category and Sport
+
 **2.0.21 (2019-03-12)**
 * exposed EventResult.getGrid()
 * exposed Reference.getAamsId()